+++ src/runtime_kqueue.h.in
@@ -1,3 +1,5 @@
+#include <sys/event.h>
+#include <fcntl.h>
 
 static void io_unpark(task_t *t);
 static int park_and_iowait(task_t **addr);
