+++ variants/gcc-linux-amd64.config
@@ -1,0 +1,9 @@
+CONFIG_CC=gcc
+CONFIG_HOSTOS=linux
+CONFIG_ARCH=amd64
+CONFIG_AR=gcc-ar
+CONFIG_POLLER=epoll
+CONFIG_INCLUDES=-I/usr/include
+CONFIG_LDFLAGS=-L/usr/lib
+CONFIG_TESTRUN=y
+CONFIG_BENCHRUN=y
