+++ variants/musl-linux-amd64.config
@@ -1,0 +1,7 @@
+CONFIG_CC=musl-gcc
+CONFIG_HOSTOS=linux
+CONFIG_AR=ar
+CONFIG_POLLER=epoll
+CONFIG_CFLAGS=-static -flto
+CONFIG_INCLUDES=-I/usr/lib/musl/include
+CONFIG_LDFLAGS=-L/usr/lib/musl/lib
