+++ variants/musl-linux-arm.config
@@ -1,0 +1,9 @@
+CONFIG_CC=musl-gcc
+CONFIG_HOSTOS=linux
+CONFIG_ARCH=arm
+CONFIG_AR=ar
+CONFIG_POLLER=epoll
+CONFIG_INCLUDES=-I/usr/include
+CONFIG_LDFLAGS=-L/usr/lib
+CONFIG_TESTRUN=y
+CONFIG_BENCHRUN=y
